Move code back to template
