+++ src/Examples.ts
@@ -3,14 +3,12 @@
  */
 import JsonFile, { JSONObject } from '@expo/json-file';
 import chalk from 'chalk';
-import execa from 'execa';
 import fs from 'fs';
 import got from 'got';
-import Minipass from 'minipass';
 import path from 'path';
 import prompts from 'prompts';
 import { Stream } from 'stream';
-import tar, { ReadEntry } from 'tar';
+import tar from 'tar';
 import terminalLink from 'terminal-link';
 import { promisify } from 'util';
 
@@ -96,7 +94,7 @@ async function isUrlOk(url: string): Promise<boolean> {
   return res.statusCode === 200;
 }
 
-export async function getRepoInfo(url: any, examplePath?: string): Promise<RepoInfo | undefined> {
+async function getRepoInfo(url: any, examplePath?: string): Promise<RepoInfo | undefined> {
   const [, username, name, t, _branch, ...file] = url.pathname.split('/');
   const filePath = examplePath ? examplePath.replace(/^\//, '') : file.join('/');
 
@@ -124,7 +122,7 @@ export async function getRepoInfo(url: any, examplePath?: string): Promise<RepoI
   return undefined;
 }
 
-export function hasRepo({ username, name, branch, filePath }: RepoInfo) {
+function hasRepo({ username, name, branch, filePath }: RepoInfo) {
   const contentsUrl = `https://api.github.com/repos/${username}/${name}/contents`;
   const packagePath = `${filePath ? `/${filePath}` : ''}/package.json`;
 
@@ -271,95 +269,6 @@ function ensureProjectHasGitIgnore(projectRoot: string): void {
 function hasExample(name: string): Promise<boolean> {
   return isUrlOk(
     `https://api.github.com/repos/expo/examples/contents/${encodeURIComponent(name)}/package.json`
-  );
-}
-
-async function getNpmUrlAsync(packageName: string): Promise<string> {
-  const url = (await execa('npm', ['v', packageName, 'dist.tarball'])).stdout;
-
-  if (!url) {
-    throw new Error(`Could not get NPM url for package "${packageName}"`);
-  }
-
-  return url;
-}
-function sanitizedName(name: string) {
-  return name
-    .replace(/[\W_]+/g, '')
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '');
-}
-
-class Transformer extends Minipass {
-  data: string;
-
-  constructor(private name: string) {
-    super();
-    this.data = '';
-  }
-  write(data: string) {
-    this.data += data;
-    return true;
-  }
-  end() {
-    let replaced = this.data
-      .replace(/Hello App Display Name/g, this.name)
-      .replace(/HelloWorld/g, sanitizedName(this.name))
-      .replace(/helloworld/g, sanitizedName(this.name.toLowerCase()));
-    super.write(replaced);
-    return super.end();
-  }
-}
-
-function createFileTransform(name: string) {
-  return (entry: ReadEntry) => {
-    // Binary files, don't process these (avoid decoding as utf8)
-    if (!['.png', '.jar', '.keystore'].includes(path.extname(entry.path)) && name) {
-      return new Transformer(name);
-    }
-    return undefined;
-  };
-}
-
-function createEntryResolver(name: string) {
-  return (entry: ReadEntry) => {
-    if (name) {
-      // Rewrite paths for bare workflow
-      entry.path = entry.path
-        .replace(
-          /HelloWorld/g,
-          entry.path.includes('android') ? sanitizedName(name.toLowerCase()) : sanitizedName(name)
-        )
-        .replace(/helloworld/g, sanitizedName(name).toLowerCase());
-    }
-    if (entry.type && /^file$/i.test(entry.type) && path.basename(entry.path) === 'gitignore') {
-      // Rename `gitignore` because npm ignores files named `.gitignore` when publishing.
-      // See: https://github.com/npm/npm/issues/1862
-      entry.path = entry.path.replace(/gitignore$/, '.gitignore');
-    }
-  };
-}
-
-export async function downloadAndExtractNpmModule(
-  root: string,
-  npmName: string,
-  projectName: string
-): Promise<void> {
-  const url = await getNpmUrlAsync(npmName);
-
-  return pipeline(
-    got.stream(url),
-    tar.extract(
-      {
-        cwd: root,
-        // TODO(ville): pending https://github.com/DefinitelyTyped/DefinitelyTyped/pull/36598
-        // @ts-ignore property missing from the type definition
-        transform: createFileTransform(projectName),
-        onentry: createEntryResolver(projectName),
-        strip: 1,
-      },
-      []
-    )
   );
 }
